+++ docs/changes.rst
@@ -8,6 +8,7 @@ To be released.
 
 - Added :class:`sider.types.Date` type.
 - Added :class:`sider.datetime.FixedOffset` tzinfo subtype.
+- Added :class:`sider.types.DateTime` type.
 
 
 Version 0.1.0